Run fewer math tests
