--- test/core/test_getgep.out
@@ -1,1 +1,0 @@
-*1 2*
